# Optimization `ChunkNoiseFunctionComponent`