--- a/pumpkin-world/Cargo.toml
+++ b/pumpkin-world/Cargo.toml
@@ -70,3 +70,7 @@ harness = false
 [[bench]]
 name = "chunk_gen"
 harness = false
+
+[[bench]]
+name = "noise_router"
+harness = false